Task 6: break circular import in transformers
Extract transform() to src/transformers/transform.ts, and the mutable
registry to src/transformers/registry.ts. index.ts populates the
registry via Object.assign at load time. All transformer files now
import {transform} from '../transform' instead of from '..', breaking
the cycle that had import-x/no-cycle disabled.

Add test/setup.ts to ensure the registry is populated before any
vitest test runs (individual transformer tests don't go through
index.ts, so a setup file is needed).

Re-enable import-x/no-cycle in xo.config.cjs.
