+++ src/transformers/input/option-group.tsx
@@ -4,7 +4,7 @@ import Text from '../../components/block/text';
 import {type Props as OptionGroupProperties} from '../../components/input/option-group';
 import {type Element} from '../../constants/types';
 import {type TextType} from '../block/text';
-import {transform} from '..';
+import {transform} from '../transform';
 import {warnIfTooLong} from '../../utils/validation';
 
 import {type OptionType} from './option';
