+++ src/transformers/input/select.tsx
@@ -6,7 +6,7 @@ import {type ConfirmationType} from '../block/confirmation';
 import {type Props as SelectProperties, selectTypes} from '../../components/input/select';
 import type Option from '../../components/input/option';
 import Text from '../../components/block/text';
-import {transform} from '..';
+import {transform} from '../transform';
 import getType from '../../utils/get-type';
 import type OptionGroup from '../../components/input/option-group';
 import {warnIfTooLong} from '../../utils/validation';
