+++ src/transformers/input/time-picker.tsx
@@ -4,7 +4,7 @@ import {type Element} from '../../constants/types';
 import {type TextType} from '../block/text';
 import {type ConfirmationType} from '../block/confirmation';
 import {type Props as TimePickerProperties} from '../../components/input/time-picker';
-import {transform} from '..';
+import {transform} from '../transform';
 import Text from '../../components/block/text';
 import {warnIfTooLong} from '../../utils/validation';
