+++ src/transformers/layout/header.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import {type Element} from '../../constants/types';
 import {type Props as HeaderProperties} from '../../components/layout/header';
 import {type TextType} from '../block/text';
-import {transform} from '..';
+import {transform} from '../transform';
 import Text from '../../components/block/text';
 import {warnIfTooLong} from '../../utils/validation';
