+++ src/transformers/layout/video.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import {type Element} from '../../constants/types';
 import {type Props as VideoProperties} from '../../components/layout/video';
 import {type TextType} from '../block/text';
-import {transform} from '..';
+import {transform} from '../transform';
 import Text from '../../components/block/text';
 import {warnIfTooLong} from '../../utils/validation';
